+++ TestErrors.swift
@@ -1,0 +1,13 @@
+//
+//  TestErrors.swift
+//  RxSwiftExt-iOS
+//
+//  Created by Anton Nazarov on 25/05/2019.
+//  Copyright © 2019 RxSwiftCommunity. All rights reserved.
+//
+
+enum TestError: Error {
+    case dummyError
+}
+
+let testError = TestError.dummyError